backend/requirements.txt
alembic==1.14.1
boto3==1.36.26
celery==5.4.0
fastapi==0.115.8
psycopg[binary]==3.2.4
PyJWT==2.10.1
pydantic-settings==2.7.1
pypdf==5.3.1
redis==5.2.1
SQLAlchemy==2.0.37
uvicorn==0.34.0
